refactor(kernel): move core contracts into kernel and enforce boundaries
