+++ src/ume/events/handlers/base.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Protocol, runtime_checkable
 
-from ...event import Event
-from ...graph_adapter import IGraphAdapter
+from ...kernel.events import Event
+from ...kernel.graph_adapter import IGraphAdapter
 
 
 @dataclass(frozen=True)
