+++ src/ume/kernel/__init__.py
@@ -1,19 +1,27 @@
-"""Kernel package exports for stable core boundaries.
+"""Kernel package exports for stable core boundaries."""
 
-The kernel contains foundational primitives shared by domain packs.
-Domain packs must depend on this layer, while kernel code must not import domains.
-"""
+from typing import Any
 
 from .events import Event
 from .graph_adapter import IGraphAdapter
-from .ledger import EventLedger
-from .policy import (
-    PolicyContext,
-    PolicyDecision,
-    PolicyPipeline,
-    build_default_policy_pipeline,
-)
+from .policy import PolicyContext, PolicyDecision, build_default_policy_pipeline
 from .processing import DEFAULT_VERSION, ProcessingError, apply_event_to_graph
+
+PolicyPipeline: Any
+EventLedger: Any
+
+
+def __getattr__(name: str):
+    if name == "PolicyPipeline":
+        from .policy import PolicyPipeline
+
+        return PolicyPipeline
+    if name == "EventLedger":
+        from .ledger import EventLedger
+
+        return EventLedger
+    raise AttributeError(name)
+
 
 __all__ = [
     "Event",
