+++ src/ume/kernel/graph_adapter.py
@@ -1,5 +1,373 @@
-"""Kernel graph adapter contracts."""
+# src/ume/graph_adapter.py
+from abc import ABC, abstractmethod
+from typing import Any, Dict, Optional, List, Tuple
 
-from ..graph_adapter import IGraphAdapter
+import asyncio
 
-__all__ = ["IGraphAdapter"]
+# To avoid circular dependency if ProcessingError is defined in processing.py
+# and processing.py imports IGraphAdapter.
+# We can define a base GraphError here or have adapter methods specify
+# that they can raise a generic Exception or a specific error type
+# that implementations must adhere to.
+# For now, let's assume a generic Exception in docstring, but implementations
+# like MockGraph will use ProcessingError from ume.processing.
+# Alternatively, we could forward declare ProcessingError if it's complex.
+
+
+class IGraphAdapter(ABC):
+    """
+    Abstract Base Class defining the interface for graph operations,
+    including node and edge manipulation, and graph state retrieval.
+
+    This interface allows different graph backend implementations (e.g.,
+    in-memory mock, database-backed) to be used interchangeably by
+    the event processing logic.
+    """
+
+    @abstractmethod
+    def add_node(self, node_id: str, attributes: Dict[str, Any]) -> None:
+        """
+        Adds a new node to the graph.
+
+        Args:
+            node_id: The unique identifier for the node.
+            attributes: A dictionary of attributes for the node.
+
+        Raises:
+            ProcessingError (or similar): If the node_id already exists.
+        """
+        pass
+
+    @abstractmethod
+    def update_node(self, node_id: str, attributes: Dict[str, Any]) -> None:
+        """
+        Updates attributes of an existing node.
+
+        Args:
+            node_id: The unique identifier for the node to update.
+            attributes: A dictionary of attributes to update.
+                        Existing attributes will be updated; new attributes
+                        will be added.
+
+        Raises:
+            ProcessingError (or similar): If the node_id does not exist.
+        """
+        pass
+
+    @abstractmethod
+    def get_node(self, node_id: str) -> Optional[Dict[str, Any]]:
+        """
+        Retrieves a node and its attributes from the graph.
+
+        Args:
+            node_id: The unique identifier for the node.
+
+        Returns:
+            A dictionary of the node's attributes if the node exists,
+            otherwise None.
+        """
+        pass
+
+    @abstractmethod
+    def node_exists(self, node_id: str) -> bool:
+        """
+        Checks if a node exists in the graph.
+
+        Args:
+            node_id: The unique identifier for the node.
+
+        Returns:
+            True if the node exists, False otherwise.
+        """
+        pass
+
+    @abstractmethod
+    def dump(self) -> Dict[str, Any]:
+        """
+        Returns a dictionary representation of the graph's current state.
+
+        This is primarily for serialization or debugging. The exact structure
+        (e.g., inclusion of nodes, edges) may depend on the implementation.
+        For consistency, it's recommended to return a dict with keys like
+        "nodes" and "edges".
+
+        Returns:
+            A dictionary representing the graph state.
+        """
+        pass
+
+    @abstractmethod
+    def clear(self) -> None:
+        """
+        Removes all nodes and edges from the graph, resetting it to an empty state.
+        """
+        pass
+
+    @abstractmethod
+    def get_all_node_ids(self) -> list[str]:
+        """
+        Retrieves a list of all node identifiers currently in the graph.
+
+        Returns:
+            A list of strings, where each string is a unique node ID.
+            Returns an empty list if the graph contains no nodes.
+        """
+        pass
+
+    @abstractmethod
+    def find_connected_nodes(
+        self, node_id: str, edge_label: Optional[str] = None
+    ) -> list[str]:
+        """
+        Finds nodes connected to a given node, optionally via a specific edge label.
+
+        This method is a placeholder for basic connectivity queries. Implementations
+        may vary based on their support for edges and edge labels.
+
+        Args:
+            node_id: The identifier of the starting node.
+            edge_label: (Optional) If provided, filter connections by this edge label.
+                        If None, consider all connections.
+
+        Returns:
+            A list of node IDs connected to the given node.
+            Returns an empty list if no connected nodes are found or if the
+            starting node does not exist or if edges are not supported.
+
+        Raises:
+            ProcessingError (or similar): If the specified node_id does not exist in the graph
+                                        (implementations may choose this behavior).
+        """
+        pass
+
+    @abstractmethod
+    def add_edge(
+        self,
+        source_node_id: str,
+        target_node_id: str,
+        label: str,
+        attrs: Dict[str, Any] | None = None,
+        schema_version: str | None = None,
+    ) -> None:
+        """
+        Adds a directed, labeled edge between two existing nodes.
+
+        Args:
+            source_node_id: The identifier of the source node (origin of the edge).
+            target_node_id: The identifier of the target node (destination of the edge).
+            label: A string label describing the type of relationship or connection.
+            attrs: Optional dictionary of edge attributes.
+            schema_version: Optional schema version for the edge definition.
+
+        Raises:
+            ProcessingError (or similar): If either the source_node_id or
+                                        target_node_id does not exist in the graph.
+                                        Implementations may also raise errors for
+                                        invalid labels or other constraints.
+        """
+        pass
+
+    @abstractmethod
+    def get_all_edges(self) -> list[tuple[str, str, str, Dict[str, Any]]]:
+        """
+        Retrieves a list of all edges currently in the graph.
+
+        Each edge is represented as a tuple:
+            (source_node_id, target_node_id, label, attrs).
+
+        Returns:
+            A list of tuples, where each tuple represents an edge and its
+            attribute dictionary. Returns an empty list if the graph contains no
+            edges.
+        """
+        pass
+
+    @abstractmethod
+    def delete_edge(
+        self,
+        source_node_id: str,
+        target_node_id: str,
+        label: str,
+        attrs: Dict[str, Any] | None = None,
+    ) -> None:
+        """
+        Removes a specific directed, labeled edge from the graph.
+
+        Args:
+            source_node_id: The identifier of the source node of the edge.
+            target_node_id: The identifier of the target node of the edge.
+            label: The label of the edge to remove.
+            attrs: Optional dictionary of edge attributes.
+
+        Raises:
+            ProcessingError (or similar): If the specified edge does not exist,
+                                        or if nodes do not exist (implementation dependent).
+        """
+        pass
+
+    @abstractmethod
+    def redact_node(self, node_id: str) -> None:
+        """Mark the specified node as redacted so it no longer appears in queries."""
+        pass
+
+    @abstractmethod
+    def redact_edge(self, source_node_id: str, target_node_id: str, label: str) -> None:
+        """Mark the specified edge as redacted so it no longer appears in queries."""
+        pass
+
+    @abstractmethod
+    def close(self) -> None:
+        """Release any resources held by the graph adapter."""
+        pass
+
+    # ---- Traversal and pathfinding -------------------------------------
+
+    @abstractmethod
+    def shortest_path(self, source_id: str, target_id: str) -> list[str]:
+        """Return the shortest path from ``source_id`` to ``target_id``.
+
+        Implementations should return an empty list if no path exists.
+        """
+        pass
+
+    @abstractmethod
+    def traverse(
+        self,
+        start_node_id: str,
+        depth: int,
+        edge_label: Optional[str] = None,
+    ) -> list[str]:
+        """Traverse outward from ``start_node_id`` up to ``depth`` hops.
+
+        ``edge_label`` may be used to restrict traversal to edges with the
+        given label. The return value is a list of visited node IDs in BFS
+        order.
+        """
+        pass
+
+    @abstractmethod
+    def extract_subgraph(
+        self,
+        start_node_id: str,
+        depth: int,
+        edge_label: Optional[str] = None,
+        since_timestamp: Optional[int] = None,
+    ) -> Dict[str, Any]:
+        """Return a dictionary describing a subgraph rooted at ``start_node_id``.
+
+        ``since_timestamp`` filters nodes to those whose ``timestamp`` attribute
+        is greater than or equal to the given epoch integer.
+        The returned dictionary must contain ``nodes`` and ``edges`` keys using
+        the same structure as :meth:`dump`.
+        """
+        pass
+
+    @abstractmethod
+    def constrained_path(
+        self,
+        source_id: str,
+        target_id: str,
+        max_depth: Optional[int] = None,
+        edge_label: Optional[str] = None,
+        since_timestamp: Optional[int] = None,
+    ) -> list[str]:
+        """Return a path from ``source_id`` to ``target_id`` using constraints.
+
+        If ``max_depth`` is provided the search will stop after that many hops.
+        ``edge_label`` restricts traversal to edges with the given label and
+        ``since_timestamp`` requires all visited nodes to have a ``timestamp``
+        attribute greater than or equal to the provided value.  Implementations
+        must return an empty list when no path satisfies the constraints.
+        """
+        pass
+
+
+class AsyncAdapterMixin:
+    """Async wrappers for CRUD operations on a synchronous :class:`IGraphAdapter`."""
+
+    _adapter: IGraphAdapter
+
+    def __init__(self, adapter: IGraphAdapter) -> None:
+        self._adapter = adapter
+
+    async def add_node(self, node_id: str, attributes: Dict[str, Any]) -> None:
+        await asyncio.to_thread(self._adapter.add_node, node_id, attributes)
+
+    async def update_node(self, node_id: str, attributes: Dict[str, Any]) -> None:
+        await asyncio.to_thread(self._adapter.update_node, node_id, attributes)
+
+    async def get_node(self, node_id: str) -> Optional[Dict[str, Any]]:
+        return await asyncio.to_thread(self._adapter.get_node, node_id)
+
+    async def node_exists(self, node_id: str) -> bool:
+        return await asyncio.to_thread(self._adapter.node_exists, node_id)
+
+    async def dump(self) -> Dict[str, Any]:
+        return await asyncio.to_thread(self._adapter.dump)
+
+    async def clear(self) -> None:
+        await asyncio.to_thread(self._adapter.clear)
+
+    async def get_all_node_ids(self) -> List[str]:
+        return await asyncio.to_thread(self._adapter.get_all_node_ids)
+
+    async def find_connected_nodes(
+        self, node_id: str, edge_label: Optional[str] = None
+    ) -> List[str]:
+        return await asyncio.to_thread(
+            self._adapter.find_connected_nodes, node_id, edge_label
+        )
+
+    async def add_edge(
+        self,
+        source_node_id: str,
+        target_node_id: str,
+        label: str,
+        attrs: Dict[str, Any] | None = None,
+        schema_version: str | None = None,
+    ) -> None:
+        await asyncio.to_thread(
+            self._adapter.add_edge,
+            source_node_id,
+            target_node_id,
+            label,
+            attrs,
+            schema_version,
+        )
+
+
+    async def get_all_edges(self) -> List[Tuple[str, str, str, Dict[str, Any]]]:
+        return await asyncio.to_thread(self._adapter.get_all_edges)
+
+    async def delete_edge(
+        self,
+        source_node_id: str,
+        target_node_id: str,
+        label: str,
+        attrs: Dict[str, Any] | None = None,
+    ) -> None:
+        if attrs is not None:
+            await asyncio.to_thread(
+                self._adapter.delete_edge,
+                source_node_id,
+                target_node_id,
+                label,
+                attrs,
+            )
+        else:
+            await asyncio.to_thread(
+                self._adapter.delete_edge, source_node_id, target_node_id, label
+            )
+
+    async def redact_node(self, node_id: str) -> None:
+        await asyncio.to_thread(self._adapter.redact_node, node_id)
+
+    async def redact_edge(
+        self, source_node_id: str, target_node_id: str, label: str
+    ) -> None:
+        await asyncio.to_thread(
+            self._adapter.redact_edge, source_node_id, target_node_id, label
+        )
+
+    async def close(self) -> None:
+        await asyncio.to_thread(self._adapter.close)
