+++ src/ume/services/mutate.py
@@ -7,21 +7,21 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Any, Callable
 
-from ..event import EventError
+from ..kernel.events import EventError
 from ..event_ledger import event_ledger
 from ..domains.classification import apply_classification
 from ..domains.extensions import DomainExtension, run_domain_extensions
 from ..events.ingress import IngressAdapter
 from ..events.schema_resolution import resolve_active_schema
 from ..config import settings
-from ..graph_adapter import IGraphAdapter
+from ..kernel.graph_adapter import IGraphAdapter
 from ..pipeline.core import (
     PipelineEnvelope,
     PipelineOutcome,
 )
-from ..policy.pipeline import PolicyDecision
+from ..kernel.policy import PolicyDecision
 from ..policy.graph_view import build_graph_read_view
-from ..processing import DEFAULT_VERSION, apply_event_to_graph
+from ..kernel.processing import DEFAULT_VERSION, apply_event_to_graph
 from ..vector_outbox import enqueue_vector_outbox_event
 from ..deprecations import warn_deprecated
