+++ tests/architecture/test_dependency_boundaries.py
@@ -1,6 +1,6 @@
 from scripts.lint_dependencies import lint_kernel_dependencies
 
 
-def test_kernel_does_not_depend_on_domains() -> None:
+def test_kernel_only_depends_on_kernel_or_external_modules() -> None:
     violations = lint_kernel_dependencies()
     assert not violations, "\n".join(violations)
